// daisy_scout_web/db/models/team-match-2025-performance.tsx
export type TeamMatch2025Performance = {
  //eventId: string;
  teamNumber: number;
  numMatches: number;
  autoStartLocAlliance: number;
  autoStartLocCenter: number;
  autoStartLocOpponent: number;
  autoStartLocNoShow: number;
  autoReefL1: number;
  autoReefL2: number;
  autoReefL3: number;
  autoReefL4: number;
  autoReefLevels: string;
  coralA: number;
  coralB: number;
  coralC: number;
  coralD: number;
  coralE: number;
  coralF: number;
  autoReefSides: string;
  leaveNum: number;
  autoNet: number;
  autoProcessor: number;
  teleopReefL1: number;
  teleopReefL2: number;
  teleopReefL3: number;
  teleopReefL4: number;
  teleopReefLevels: string;
  teleopProcessor: number;
  teleopNet: number;
  teleopAlgaeRemoved: number;
  endgamePark: number;
  endgameShallow: number;
  endgameDeep: number;
  endgameNone: number;
  endgame: string;
  driverAbility: number | null | undefined;
  operability: string | null | undefined;
  role: string | null | undefined;
  hpNetShots: string | null | undefined;
  hpNetMade: number | null | undefined;
};