// daisy_scout_web/db/models/sharable-match-2025-data.tsx
export type SharableMatch2025Data = {
  teamNumber: number;
  asOfMatch: number | null;
  totalCoral: number;

  totalAutoCoral: number;
  autoReefL1: string | null;
  autoReefL2: string | null;
  autoReefL3: string | null;
  autoReefL4: string | null;
  leave: string | null;
  autoNet: string | null;
  autoProcessor: string | null;
  totalTeleopCoral: number;
  teleopReefL1: string | null;
  teleopReefL2: string | null;
  teleopReefL3: string | null;
  teleopReefL4: string | null;
  teleopProcessor: string | null;
  teleopNet: string | null;
  teleopAlgaeRemoved: string | null;
};